requirement.txt
allure-behave==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==2.1.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
bdd==0.2
behave==1.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==36.0.0
enum34==1.1.10
future==0.18.2
glob2==0.7
h11==0.12.0
idna==3.3
iniconfig==1.1.1
lxml==4.7.1
Mako==1.1.6
MarkupSafe==2.0.1
namedlist==1.8
outcome==1.1.0
packaging==21.3
parse==1.19.0
parse-type==0.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
pytest-bdd==5.0.0
pytest-html==3.1.1
pytest-metadata==1.11.0
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webdriver-manager==3.5.2
wsproto==1.0.0
